fix up review comments
